[MOV] core: service.model.retrying -> http.retrying
