+++ addons/mail/models/mail_presence.py
@@ -3,7 +3,7 @@
 from datetime import timedelta
 
 from odoo import api, fields, models, tools
-from odoo.service.model import PG_CONCURRENCY_EXCEPTIONS_TO_RETRY
+from odoo.sql_db import PG_CONCURRENCY_EXCEPTIONS_TO_RETRY
 
 UPDATE_PRESENCE_DELAY = 60
 DISCONNECTION_TIMER = UPDATE_PRESENCE_DELAY + 5
